separate RTC & Time classes also added global_RTC
